business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/SendMorse.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class SendMorse implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+b ");
        messageString.append(command.getParameter("Morse"));
        messageString.append("\n");
        return messageString;
    }
}
